Update Crates (#131)
* move all deps to workspace, remove patch ver

* update deps

* updating axum pain

* more axum updates pain

* figured out a cleaner approach to fix axum

* more examples

* finish updating axum in a clean manner

* doc comments, clean up, and fix mentat macro

* clippy fixes

+++ crates/mentat-asserter/src/network.rs
@@ -97,7 +97,7 @@ pub fn network_status_response(resp: Option<&UncheckedNetworkStatusResponse>) ->
             resp.current_block_identifier
         )
     })?;
-    timestamp(resp.current_block_timestamp as isize).map_err(|e| {
+    timestamp(resp.current_block_timestamp).map_err(|e| {
         format!(
             "current block timestamp {} is invalid: {e}",
             resp.current_block_timestamp
